chore(data): move Config.java to data
