+++ app/src/main/java/com/osfans/trime/ime/keyboard/Event.java
@@ -22,7 +22,7 @@ import android.text.TextUtils;
 import android.view.KeyEvent;
 import androidx.annotation.NonNull;
 import com.osfans.trime.core.Rime;
-import com.osfans.trime.setup.Config;
+import com.osfans.trime.data.Config;
 import com.osfans.trime.util.ConfigGetter;
 import java.util.HashMap;
 import java.util.List;
