+++ app/src/main/java/com/osfans/trime/ime/keyboard/Key.java
@@ -26,7 +26,7 @@ import android.view.KeyCharacterMap;
 import android.view.KeyEvent;
 import com.osfans.trime.core.Rime;
 import com.osfans.trime.ime.enums.KeyEventType;
-import com.osfans.trime.setup.Config;
+import com.osfans.trime.data.Config;
 import com.osfans.trime.util.ConfigGetter;
 import java.util.List;
 import java.util.Map;
