+++ app/src/main/java/com/osfans/trime/ime/keyboard/KeyboardSwitcher.kt
@@ -2,7 +2,7 @@ package com.osfans.trime.ime.keyboard
 
 import android.content.res.Configuration
 import com.osfans.trime.ime.core.Trime
-import com.osfans.trime.setup.Config
+import com.osfans.trime.data.Config
 import timber.log.Timber
 
 /** Manages [Keyboard]s and their status. **/
