+++ app/src/main/java/com/osfans/trime/ime/symbol/SimpleAdapter.java
@@ -14,7 +14,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 import com.google.android.flexbox.FlexboxLayoutManager;
 import com.osfans.trime.R;
-import com.osfans.trime.setup.Config;
+import com.osfans.trime.data.Config;
 import java.util.List;
 import timber.log.Timber;
