+++ app/src/main/java/com/osfans/trime/ime/text/TextInputManager.kt
@@ -17,7 +17,7 @@ import com.osfans.trime.ime.keyboard.Key
 import com.osfans.trime.ime.keyboard.Keyboard.printModifierKeyState
 import com.osfans.trime.ime.keyboard.KeyboardSwitcher
 import com.osfans.trime.ime.keyboard.KeyboardView
-import com.osfans.trime.setup.Config
+import com.osfans.trime.data.Config
 import com.osfans.trime.setup.IntentReceiver
 import com.osfans.trime.util.ShortcutUtils
 import kotlinx.coroutines.CoroutineScope
