+++ app/src/main/java/com/osfans/trime/settings/components/ColorPickerDialog.kt
@@ -7,7 +7,7 @@ import androidx.appcompat.app.AlertDialog
 import com.osfans.trime.R
 import com.osfans.trime.ime.core.Preferences
 import com.osfans.trime.ime.core.Trime
-import com.osfans.trime.setup.Config
+import com.osfans.trime.data.Config
 import timber.log.Timber
 
 /** 顯示配色方案列表
