+++ app/src/main/java/com/osfans/trime/settings/components/ThemePickerDialog.kt
@@ -8,7 +8,7 @@ import android.view.WindowManager
 import androidx.appcompat.app.AlertDialog
 import com.osfans.trime.R
 import com.osfans.trime.ime.core.Trime
-import com.osfans.trime.setup.Config
+import com.osfans.trime.data.Config
 import com.osfans.trime.util.createLoadingDialog
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
